Update checkAuth.js
Reverting back to user instead of userSafe. Logic breaks calls and makes it so we have to call (example) userData.user.firstname instead of userData.firstname because, for this logic to work properly, the renaming would need to say {user: safeUser} in the 200 response message instead for all calls that use this.